feat(output): Accept passed in package set in standalone mode
Sometimes it may be desirable to have a standalone derivation (i.e. a
derivation that can be built simply by calling `nix-build`)
that *also* lets the user pass in a package set if desired.

This updates the preamble written in standalone-mode to accept a
`pkgs` argument which defaults to the globally importable package set
if not provided.

By default an expression written this way will not behave differently
from one that uses `with import <nixpkgs> {}`.